Moved module to right directory and created symlink.
